# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.2",
    "title": "IP Snapshot",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "",
...
  "schemes": ["http"],
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["user"],
        "summary": "List of all Users",
        "description": "List of all Users",
        "operationId": "getUsers",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "post": {
        "tags": ["user"],
        "summary": " New User",
        "description": "Add a new user into db",
        "operationId": "addUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User that needs to be added to the store",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
    "/api/users/{id}": {
      "delete": {
        "tags": ["user"],
        "summary": "Delete user by ID(set active=0)",
        "description": "",
        "operationId": "deleteUser",
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user that needs to be deleted",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "patch": {
        "tags": ["user"],
        "summary": " Edit user",
        "description": "Edit user information into db",
        "operationId": "patchUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user that needs to be modified",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EditUsers"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "get": {
        "tags": ["user"],
        "summary": "Get Users by Id",
        "description": "List of Users data with departments & Permissions",
        "operationId": "getAllUserData Matrix",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user for getting all the information.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Users": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "departmentId": {
          "type": "integer"
        }
      }
    },
    "EditUsers": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "departmentId": {
          "type": "integer"
        },
        "permissionId": {
          "type": "integer"
        },
        "isActive": {
          "type": "integer"
        }
      }
    },
...
  },
...
}